pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cookimport"
version = "0.1.0"
description = "Local CLI for importing recipe sources into staging JSON-LD."
readme = "cookimport/README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12",
  "ebooklib>=0.17",
  "ingredient-parser-nlp>=0.1.0",
  "lxml>=5.0",
  "markitdown>=0.1.4,<0.2",
  "markdownify>=1.2,<2",
  "openpyxl>=3.1",
  "pymupdf>=1.23",
  "pydantic>=2.5,<3",
  "PyYAML>=6.0",
  "questionary>=2.0",
  "rapidfuzz>=3.0",
  "rich>=13.7",
  "typer>=0.12",
  "python-docx>=1.1.0",
  "python-doctr[torch]>=0.9",
  "unstructured>=0.18.32,<0.19",
]

[project.optional-dependencies]
db = [
  "psycopg[binary]>=3.1",
]
benchaccel = [
  "cydifflib",
  "cdifflib",
  "fast-diff-match-patch",
]
dev = [
  "pytest>=7.4",
  "psycopg[binary]>=3.1",
]
segmentation_eval = [
  "segeval>=2.0.11",
]
epubdebug = [
  "epub-utils==0.1.0a1",
]
webschema = [
  "boilerpy3>=1.0.7",
  "extruct>=0.18",
  "isodate>=0.6",
  "justext>=3.0",
  "pyld>=2.0",
  "readability-lxml>=0.8.4",
  "recipe-scrapers>=15.9",
  "scrape-schema-recipe",
  "trafilatura>=1.8",
]
ingredientparse = [
  "ftfy>=6.2",
  "quantulum3>=0.9",
  "pint>=0.23",
]

[project.scripts]
cookimport = "cookimport.cli:app"
import = "cookimport.entrypoint:main"
C3import = "cookimport.entrypoint:main"
C3imp = "cookimport.c3imp_entrypoint:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cookimport*"]
